apps/ui/requirements.txt
streamlit==1.35.0
requests==2.32.5
python-dotenv==1.0.1
streamlit-audiorecorder==0.0.5